Rename each main component file to index, to improve imports
