--- src/Containers/Calculator/index.js
@@ -1,3 +1,0 @@
-import Calculator from './Calculator';
-
-export default Calculator;
