--- src/Containers/Navbar/index.js
@@ -1,3 +1,0 @@
-import Navbar from './Navbar';
-
-export default Navbar;
